+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.63.0]
+
+* When using the "Save All" command, call the "Logtalk: Make - Reload" command once after all files are saved rather than per file
+
 ## [0.62.0]
 
 * Improve predicate declarations formatting when predicates are also operators